tests infoclimat

+++ infoclimat.go
@@ -1,7 +1,7 @@
 package main
 
 type InfoClimatResponse struct {
-	RequestState int              `json:"request_state"`
+	RequestState int16            `json:"request_state"`
 	RequestKey   string           `json:"request_key"`
 	Message      string           `json:"message"`
 	ModelRun     string           `json:"model_run"`
@@ -11,20 +11,20 @@ type InfoClimatResponse struct {
 
 type Meteo struct {
 	Temperature struct {
-		TwoM       float64 `json:"2m"`
-		Sol        float64 `json:"sol"`
-		Five00HPa  float64 `json:"500hPa"`
-		Eight50HPa float64 `json:"850hPa"`
+		TwoM       float32 `json:"2m"`
+		Sol        float32 `json:"sol"`
+		Five00HPa  float32 `json:"500hPa"`
+		Eight50HPa float32 `json:"850hPa"`
 	} `json:"temperature"`
 	Pression struct {
 		NiveauDeLaMer int `json:"niveau_de_la_mer"`
 	} `json:"pression"`
-	Pluie    float64 `json:"pluie"`
+	Pluie    float32 `json:"pluie"`
 	Humidite struct {
-		TwoM int `json:"2m"`
+		TwoM float32 `json:"2m"`
 	} `json:"humidite"`
 	VentMoyen struct {
-		One0M float64 `json:"10m"`
+		One0M float32 `json:"10m"`
 	} `json:"vent_moyen"`
 	RisqueNeige string `json:"risque_neige"`
 	Nebulosite  struct {
